Fixed auth
